% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Account}
\alias{Account}
\title{Tag Manager API Objects 
Accesses Tag Manager accounts and containers.}
\usage{
Account(accountId = NULL, fingerprint = NULL, name = NULL,
  shareData = NULL)
}
\arguments{
\item{accountId}{The Account ID uniquely identifies the GTM Account}

\item{fingerprint}{The fingerprint of the GTM Account as computed at storage time}

\item{name}{Account display name}

\item{shareData}{Whether the account shares data anonymously with Google and others}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:18:35
filename: /Users/mark/dev/R/autoGoogleAPI/googletagmanagerv1.auto/R/tagmanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Account.
}
\seealso{
Other Account functions: \code{\link{accounts.update}}
}
