% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{AccountAccess}
\alias{AccountAccess}
\title{AccountAccess Object}
\usage{
AccountAccess(permission = NULL)
}
\arguments{
\item{permission}{List of Account permissions}
}
\value{
AccountAccess object
}
\description{
AccountAccess Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the Google Tag Manager Account access permissions.
}

