% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_functions.R
\name{accounts.get}
\alias{accounts.get}
\title{Gets a GTM Account.}
\usage{
accounts.get(accountId)
}
\arguments{
\item{accountId}{The GTM Account ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tagmanager.edit.containers
\item https://www.googleapis.com/auth/tagmanager.manage.accounts
\item https://www.googleapis.com/auth/tagmanager.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tagmanager.edit.containers, https://www.googleapis.com/auth/tagmanager.manage.accounts, https://www.googleapis.com/auth/tagmanager.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/tag-manager/api/v1/}{Google Documentation}
}

