% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_functions.R
\name{accounts.update}
\alias{accounts.update}
\title{Updates a GTM Account.}
\usage{
accounts.update(Account, accountId, fingerprint = NULL)
}
\arguments{
\item{Account}{The \link{Account} object to pass to this method}

\item{accountId}{The GTM Account ID}

\item{fingerprint}{When provided, this fingerprint must match the fingerprint of the account in storage}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/tagmanager.manage.accounts
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tagmanager.manage.accounts)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/tag-manager/api/v1/}{Google Documentation}

Other Account functions: \code{\link{Account}}
}

