% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Condition}
\alias{Condition}
\title{Condition Object}
\usage{
Condition(parameter = NULL, type = NULL)
}
\arguments{
\item{parameter}{A list of named parameters (key/value), depending on the condition's type}

\item{type}{The type of operator for this condition}
}
\value{
Condition object
}
\description{
Condition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a predicate.
}

