% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Container}
\alias{Container}
\title{Container Object}
\usage{
Container(accountId = NULL, containerId = NULL, domainName = NULL,
  enabledBuiltInVariable = NULL, fingerprint = NULL, name = NULL,
  notes = NULL, publicId = NULL, timeZoneCountryId = NULL,
  timeZoneId = NULL, usageContext = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{containerId}{The Container ID uniquely identifies the GTM Container}

\item{domainName}{Optional list of domain names associated with the Container}

\item{enabledBuiltInVariable}{List of enabled built-in variables}

\item{fingerprint}{The fingerprint of the GTM Container as computed at storage time}

\item{name}{Container display name}

\item{notes}{Container Notes}

\item{publicId}{Container Public ID}

\item{timeZoneCountryId}{Container Country ID}

\item{timeZoneId}{Container Time Zone ID}

\item{usageContext}{List of Usage Contexts for the Container}
}
\value{
Container object
}
\description{
Container Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Container.
}

