% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ContainerAccess}
\alias{ContainerAccess}
\title{ContainerAccess Object}
\usage{
ContainerAccess(containerId = NULL, permission = NULL)
}
\arguments{
\item{containerId}{GTM Container ID}

\item{permission}{List of Container permissions}
}
\value{
ContainerAccess object
}
\description{
ContainerAccess Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines the Google Tag Manager Container access permissions.
}

