% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ContainerVersion}
\alias{ContainerVersion}
\title{ContainerVersion Object}
\usage{
ContainerVersion(accountId = NULL, container = NULL, containerId = NULL,
  containerVersionId = NULL, deleted = NULL, fingerprint = NULL,
  folder = NULL, macro = NULL, name = NULL, notes = NULL, rule = NULL,
  tag = NULL, trigger = NULL, variable = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{container}{The container that this version was taken from}

\item{containerId}{GTM Container ID}

\item{containerVersionId}{The Container Version ID uniquely identifies the GTM Container Version}

\item{deleted}{A value of true indicates this container version has been deleted}

\item{fingerprint}{The fingerprint of the GTM Container Version as computed at storage time}

\item{folder}{The folders in the container that this version was taken from}

\item{macro}{The macros in the container that this version was taken from}

\item{name}{Container version display name}

\item{notes}{User notes on how to apply this container version in the container}

\item{rule}{The rules in the container that this version was taken from}

\item{tag}{The tags in the container that this version was taken from}

\item{trigger}{The triggers in the container that this version was taken from}

\item{variable}{The variables in the container that this version was taken from}
}
\value{
ContainerVersion object
}
\description{
ContainerVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Container Version.
}

