% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ContainerVersionHeader}
\alias{ContainerVersionHeader}
\title{ContainerVersionHeader Object}
\usage{
ContainerVersionHeader(accountId = NULL, containerId = NULL,
  containerVersionId = NULL, deleted = NULL, name = NULL,
  numMacros = NULL, numRules = NULL, numTags = NULL, numTriggers = NULL,
  numVariables = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{containerId}{GTM Container ID}

\item{containerVersionId}{The Container Version ID uniquely identifies the GTM Container Version}

\item{deleted}{A value of true indicates this container version has been deleted}

\item{name}{Container version display name}

\item{numMacros}{Number of macros in the container version}

\item{numRules}{Number of rules in the container version}

\item{numTags}{Number of tags in the container version}

\item{numTriggers}{Number of triggers in the container version}

\item{numVariables}{Number of variables in the container version}
}
\value{
ContainerVersionHeader object
}
\description{
ContainerVersionHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Container Version Header.
}

