% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{CreateContainerVersionRequestVersionOptions}
\alias{CreateContainerVersionRequestVersionOptions}
\title{CreateContainerVersionRequestVersionOptions Object}
\usage{
CreateContainerVersionRequestVersionOptions(name = NULL, notes = NULL,
  quickPreview = NULL)
}
\arguments{
\item{name}{The name of the container version to be created}

\item{notes}{The notes of the container version to be created}

\item{quickPreview}{The creation of this version may be for quick preview and shouldn't be saved}
}
\value{
CreateContainerVersionRequestVersionOptions object
}
\description{
CreateContainerVersionRequestVersionOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for new container versions.
}
