% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Environment}
\alias{Environment}
\title{Environment Object}
\usage{
Environment(accountId = NULL, authorizationCode = NULL,
  authorizationTimestampMs = NULL, containerId = NULL,
  containerVersionId = NULL, description = NULL, enableDebug = NULL,
  environmentId = NULL, fingerprint = NULL, name = NULL, type = NULL,
  url = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{authorizationCode}{The environment authorization code}

\item{authorizationTimestampMs}{The last update time-stamp for the authorization code}

\item{containerId}{GTM Container ID}

\item{containerVersionId}{No description}

\item{description}{The environment description}

\item{enableDebug}{Whether or not to enable debug by default on for the environment}

\item{environmentId}{GTM Environment ID uniquely identifies the GTM Environment}

\item{fingerprint}{The fingerprint of the GTM environment as computed at storage time}

\item{name}{The environment display name}

\item{type}{The type of this environment}

\item{url}{Default preview page url for the environment}
}
\value{
Environment object
}
\description{
Environment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Environment. Note that a user can create, delete and update environments of type USER, but can only update the enable_debug and url fields of environments of other types.
}

