% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Folder}
\alias{Folder}
\title{Folder Object}
\usage{
Folder(accountId = NULL, containerId = NULL, fingerprint = NULL,
  folderId = NULL, name = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{containerId}{GTM Container ID}

\item{fingerprint}{The fingerprint of the GTM Folder as computed at storage time}

\item{folderId}{The Folder ID uniquely identifies the GTM Folder}

\item{name}{Folder display name}
}
\value{
Folder object
}
\description{
Folder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Folder.
}

