% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{FolderEntities}
\alias{FolderEntities}
\title{FolderEntities Object}
\usage{
FolderEntities(tag = NULL, trigger = NULL, variable = NULL)
}
\arguments{
\item{tag}{The list of tags inside the folder}

\item{trigger}{The list of triggers inside the folder}

\item{variable}{The list of variables inside the folder}
}
\value{
FolderEntities object
}
\description{
FolderEntities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Folder's contents.
}
