% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ListAccountsResponse}
\alias{ListAccountsResponse}
\title{ListAccountsResponse Object}
\usage{
ListAccountsResponse(accounts = NULL)
}
\arguments{
\item{accounts}{List of GTM Accounts that a user has access to}
}
\value{
ListAccountsResponse object
}
\description{
ListAccountsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List Accounts Response.
}

