% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ListFoldersResponse}
\alias{ListFoldersResponse}
\title{ListFoldersResponse Object}
\usage{
ListFoldersResponse(folders = NULL)
}
\arguments{
\item{folders}{All GTM Folders of a GTM Container}
}
\value{
ListFoldersResponse object
}
\description{
ListFoldersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List Folders Response.
}
