% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ListTriggersResponse}
\alias{ListTriggersResponse}
\title{ListTriggersResponse Object}
\usage{
ListTriggersResponse(triggers = NULL)
}
\arguments{
\item{triggers}{All GTM Triggers of a GTM Container}
}
\value{
ListTriggersResponse object
}
\description{
ListTriggersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List triggers response.
}
