% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{ListVariablesResponse}
\alias{ListVariablesResponse}
\title{ListVariablesResponse Object}
\usage{
ListVariablesResponse(variables = NULL)
}
\arguments{
\item{variables}{All GTM Variables of a GTM Container}
}
\value{
ListVariablesResponse object
}
\description{
ListVariablesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List Variables Response.
}

