% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Macro}
\alias{Macro}
\title{Macro Object}
\usage{
Macro(accountId = NULL, containerId = NULL, disablingRuleId = NULL,
  enablingRuleId = NULL, fingerprint = NULL, macroId = NULL,
  name = NULL, notes = NULL, parameter = NULL, parentFolderId = NULL,
  scheduleEndMs = NULL, scheduleStartMs = NULL, type = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{containerId}{GTM Container ID}

\item{disablingRuleId}{For mobile containers only: A list of rule IDs for disabling conditional macros; the macro is enabled if one of the enabling rules is true while all the disabling rules are false}

\item{enablingRuleId}{For mobile containers only: A list of rule IDs for enabling conditional macros; the macro is enabled if one of the enabling rules is true while all the disabling rules are false}

\item{fingerprint}{The fingerprint of the GTM Macro as computed at storage time}

\item{macroId}{The Macro ID uniquely identifies the GTM Macro}

\item{name}{Macro display name}

\item{notes}{User notes on how to apply this macro in the container}

\item{parameter}{The macro's parameters}

\item{parentFolderId}{Parent folder id}

\item{scheduleEndMs}{The end timestamp in milliseconds to schedule a macro}

\item{scheduleStartMs}{The start timestamp in milliseconds to schedule a macro}

\item{type}{GTM Macro Type}
}
\value{
Macro object
}
\description{
Macro Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Macro.
}

