% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Parameter}
\alias{Parameter}
\title{Parameter Object}
\usage{
Parameter(key = NULL, list = NULL, map = NULL, type = NULL,
  value = NULL)
}
\arguments{
\item{key}{The named key that uniquely identifies a parameter}

\item{list}{This list parameter's parameters (keys will be ignored)}

\item{map}{This map parameter's parameters (must have keys; keys must be unique)}

\item{type}{The parameter type}

\item{value}{A parameter's value (may contain variable references such as '{{myVariable}}') as appropriate to the specified type}
}
\value{
Parameter object
}
\description{
Parameter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Parameter.
}
