% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Rule}
\alias{Rule}
\title{Rule Object}
\usage{
Rule(accountId = NULL, condition = NULL, containerId = NULL,
  fingerprint = NULL, name = NULL, notes = NULL, ruleId = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{condition}{The list of conditions that make up this rule (implicit AND between them)}

\item{containerId}{GTM Container ID}

\item{fingerprint}{The fingerprint of the GTM Rule as computed at storage time}

\item{name}{Rule display name}

\item{notes}{User notes on how to apply this rule in the container}

\item{ruleId}{The Rule ID uniquely identifies the GTM Rule}
}
\value{
Rule object
}
\description{
Rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Rule.
}

