% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{SetupTag}
\alias{SetupTag}
\title{SetupTag Object}
\usage{
SetupTag(stopOnSetupFailure = NULL, tagName = NULL)
}
\arguments{
\item{stopOnSetupFailure}{If true, fire the main tag if and only if the setup tag fires successfully}

\item{tagName}{The name of the setup tag}
}
\value{
SetupTag object
}
\description{
SetupTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
