% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Tag}
\alias{Tag}
\title{Tag Object}
\usage{
Tag(accountId = NULL, blockingRuleId = NULL, blockingTriggerId = NULL,
  containerId = NULL, fingerprint = NULL, firingRuleId = NULL,
  firingTriggerId = NULL, liveOnly = NULL, name = NULL, notes = NULL,
  parameter = NULL, parentFolderId = NULL, priority = NULL,
  scheduleEndMs = NULL, scheduleStartMs = NULL, setupTag = NULL,
  tagFiringOption = NULL, tagId = NULL, teardownTag = NULL, type = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{blockingRuleId}{Blocking rule IDs}

\item{blockingTriggerId}{Blocking trigger IDs}

\item{containerId}{GTM Container ID}

\item{fingerprint}{The fingerprint of the GTM Tag as computed at storage time}

\item{firingRuleId}{Firing rule IDs}

\item{firingTriggerId}{Firing trigger IDs}

\item{liveOnly}{If set to true, this tag will only fire in the live environment (e}

\item{name}{Tag display name}

\item{notes}{User notes on how to apply this tag in the container}

\item{parameter}{The tag's parameters}

\item{parentFolderId}{Parent folder id}

\item{priority}{User defined numeric priority of the tag}

\item{scheduleEndMs}{The end timestamp in milliseconds to schedule a tag}

\item{scheduleStartMs}{The start timestamp in milliseconds to schedule a tag}

\item{setupTag}{The list of setup tags}

\item{tagFiringOption}{Option to fire this tag}

\item{tagId}{The Tag ID uniquely identifies the GTM Tag}

\item{teardownTag}{The list of teardown tags}

\item{type}{GTM Tag Type}
}
\value{
Tag object
}
\description{
Tag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Tag.
}

