% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_functions.R
\docType{package}
\name{tagmanager_googleAuthR}
\alias{tagmanager_googleAuthR}
\alias{tagmanager_googleAuthR-package}
\title{Tag Manager API
Accesses Tag Manager accounts and containers.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:48:57
filename: /Users/mark/dev/R/autoGoogleAPI/googletagmanagerv1.auto/R/tagmanager_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/tagmanager.delete.containers
\item https://www.googleapis.com/auth/tagmanager.edit.containers
\item https://www.googleapis.com/auth/tagmanager.edit.containerversions
\item https://www.googleapis.com/auth/tagmanager.manage.accounts
\item https://www.googleapis.com/auth/tagmanager.manage.users
\item https://www.googleapis.com/auth/tagmanager.publish
\item https://www.googleapis.com/auth/tagmanager.readonly
}
}

