% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{TeardownTag}
\alias{TeardownTag}
\title{TeardownTag Object}
\usage{
TeardownTag(stopTeardownOnFailure = NULL, tagName = NULL)
}
\arguments{
\item{stopTeardownOnFailure}{If true, fire the teardown tag if and only if the main tag fires successfully}

\item{tagName}{The name of the teardown tag}
}
\value{
TeardownTag object
}
\description{
TeardownTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

