% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Trigger}
\alias{Trigger}
\title{Trigger Object}
\usage{
Trigger(accountId = NULL, autoEventFilter = NULL, checkValidation = NULL,
  containerId = NULL, customEventFilter = NULL, enableAllVideos = NULL,
  eventName = NULL, filter = NULL, fingerprint = NULL, interval = NULL,
  limit = NULL, name = NULL, parentFolderId = NULL, triggerId = NULL,
  type = NULL, uniqueTriggerId = NULL, videoPercentageList = NULL,
  waitForTags = NULL, waitForTagsTimeout = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{autoEventFilter}{Used in the case of auto event tracking}

\item{checkValidation}{Whether or not we should only fire tags if the form submit or link click event is not cancelled by some other event handler (e}

\item{containerId}{GTM Container ID}

\item{customEventFilter}{Used in the case of custom event, which is fired iff all Conditions are true}

\item{enableAllVideos}{Reloads the videos in the page that don't already have the YT API enabled}

\item{eventName}{Name of the GTM event that is fired}

\item{filter}{The trigger will only fire iff all Conditions are true}

\item{fingerprint}{The fingerprint of the GTM Trigger as computed at storage time}

\item{interval}{Time between triggering recurring Timer Events (in milliseconds)}

\item{limit}{Limit of the number of GTM events this Timer Trigger will fire}

\item{name}{Trigger display name}

\item{parentFolderId}{Parent folder id}

\item{triggerId}{The Trigger ID uniquely identifies the GTM Trigger}

\item{type}{Defines the data layer event that causes this trigger}

\item{uniqueTriggerId}{Globally unique id of the trigger that auto-generates this (a Form Submit, Link Click or Timer listener) if any}

\item{videoPercentageList}{List of integer percentage values}

\item{waitForTags}{Whether or not we should delay the form submissions or link opening until all of the tags have fired (by preventing the default action and later simulating the default action)}

\item{waitForTagsTimeout}{How long to wait (in milliseconds) for tags to fire when 'waits_for_tags' above evaluates to true}
}
\value{
Trigger object
}
\description{
Trigger Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Trigger
}

