% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{UserAccess}
\alias{UserAccess}
\title{UserAccess Object}
\usage{
UserAccess(accountAccess = NULL, accountId = NULL, containerAccess = NULL,
  emailAddress = NULL, permissionId = NULL)
}
\arguments{
\item{accountAccess}{GTM Account access permissions}

\item{accountId}{GTM Account ID}

\item{containerAccess}{GTM Container access permissions}

\item{emailAddress}{User's email address}

\item{permissionId}{Account Permission ID}
}
\value{
UserAccess object
}
\description{
UserAccess Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a user's permissions to an account and its container.
}

