% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmanager_objects.R
\name{Variable}
\alias{Variable}
\title{Variable Object}
\usage{
Variable(accountId = NULL, containerId = NULL, disablingTriggerId = NULL,
  enablingTriggerId = NULL, fingerprint = NULL, name = NULL,
  notes = NULL, parameter = NULL, parentFolderId = NULL,
  scheduleEndMs = NULL, scheduleStartMs = NULL, type = NULL,
  variableId = NULL)
}
\arguments{
\item{accountId}{GTM Account ID}

\item{containerId}{GTM Container ID}

\item{disablingTriggerId}{For mobile containers only: A list of trigger IDs for disabling conditional variables; the variable is enabled if one of the enabling trigger is true while all the disabling trigger are false}

\item{enablingTriggerId}{For mobile containers only: A list of trigger IDs for enabling conditional variables; the variable is enabled if one of the enabling triggers is true while all the disabling triggers are false}

\item{fingerprint}{The fingerprint of the GTM Variable as computed at storage time}

\item{name}{Variable display name}

\item{notes}{User notes on how to apply this variable in the container}

\item{parameter}{The variable's parameters}

\item{parentFolderId}{Parent folder id}

\item{scheduleEndMs}{The end timestamp in milliseconds to schedule a variable}

\item{scheduleStartMs}{The start timestamp in milliseconds to schedule a variable}

\item{type}{GTM Variable Type}

\item{variableId}{The Variable ID uniquely identifies the GTM Variable}
}
\value{
Variable object
}
\description{
Variable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Google Tag Manager Variable.
}

